requirements.txt
flask>=2.0
transformers>=4.0
Pillow>=8.0
requests
